fix(test): use int64 types in BoundedCounter test assertions
The BoundedCounter.Get() method returns int64, but tests were using
int literals in assert.Equal() which caused type mismatch failures
on CI.

+++ internal/pkg/tui/components/bounded_counter_test.go
@@ -16,9 +16,9 @@ func TestBoundedCounter_BasicIncrement(t *testing.T) {
 	bc.Increment("foo")
 	bc.Increment("bar")
 
-	assert.Equal(t, 2, bc.Get("foo"))
-	assert.Equal(t, 1, bc.Get("bar"))
-	assert.Equal(t, 0, bc.Get("nonexistent"))
+	assert.Equal(t, int64(2), bc.Get("foo"))
+	assert.Equal(t, int64(1), bc.Get("bar"))
+	assert.Equal(t, int64(0), bc.Get("nonexistent"))
 	assert.Equal(t, 2, bc.Len())
 }
 
@@ -36,7 +36,7 @@ func TestBoundedCounter_CapacityLimit(t *testing.T) {
 	bc.Increment("d")
 
 	assert.Equal(t, 3, bc.Len(), "should maintain capacity limit")
-	assert.Equal(t, 1, bc.Get("d"), "new item should be added")
+	assert.Equal(t, int64(1), bc.Get("d"), "new item should be added")
 }
 
 func TestBoundedCounter_EvictsLowestCount(t *testing.T) {
@@ -55,10 +55,10 @@ func TestBoundedCounter_EvictsLowestCount(t *testing.T) {
 	// Add new item - should evict "low" (count=1)
 	bc.Increment("new")
 
-	assert.Equal(t, 0, bc.Get("low"), "lowest count item should be evicted")
-	assert.Equal(t, 3, bc.Get("high"))
-	assert.Equal(t, 2, bc.Get("medium"))
-	assert.Equal(t, 1, bc.Get("new"))
+	assert.Equal(t, int64(0), bc.Get("low"), "lowest count item should be evicted")
+	assert.Equal(t, int64(3), bc.Get("high"))
+	assert.Equal(t, int64(2), bc.Get("medium"))
+	assert.Equal(t, int64(1), bc.Get("new"))
 }
 
 func TestBoundedCounter_IncrementExistingDoesNotEvict(t *testing.T) {
@@ -73,9 +73,9 @@ func TestBoundedCounter_IncrementExistingDoesNotEvict(t *testing.T) {
 	bc.Increment("a")
 
 	assert.Equal(t, 3, bc.Len())
-	assert.Equal(t, 3, bc.Get("a"))
-	assert.Equal(t, 1, bc.Get("b"))
-	assert.Equal(t, 1, bc.Get("c"))
+	assert.Equal(t, int64(3), bc.Get("a"))
+	assert.Equal(t, int64(1), bc.Get("b"))
+	assert.Equal(t, int64(1), bc.Get("c"))
 }
 
 func TestBoundedCounter_GetAll(t *testing.T) {
@@ -87,13 +87,13 @@ func TestBoundedCounter_GetAll(t *testing.T) {
 
 	all := bc.GetAll()
 
-	assert.Equal(t, 2, all["foo"])
-	assert.Equal(t, 1, all["bar"])
+	assert.Equal(t, int64(2), all["foo"])
+	assert.Equal(t, int64(1), all["bar"])
 	assert.Equal(t, 2, len(all))
 
 	// Verify it's a copy - modifying result doesn't affect counter
 	all["foo"] = 100
-	assert.Equal(t, 2, bc.Get("foo"))
+	assert.Equal(t, int64(2), bc.Get("foo"))
 }
 
 func TestBoundedCounter_GetTopN(t *testing.T) {
@@ -113,9 +113,9 @@ func TestBoundedCounter_GetTopN(t *testing.T) {
 
 	assert.Len(t, top2, 2)
 	assert.Equal(t, "high", top2[0].Key)
-	assert.Equal(t, 10, top2[0].Count)
+	assert.Equal(t, int64(10), top2[0].Count)
 	assert.Equal(t, "medium", top2[1].Key)
-	assert.Equal(t, 5, top2[1].Count)
+	assert.Equal(t, int64(5), top2[1].Count)
 
 	// Get top 10 (more than available) - should return all 3
 	top10 := bc.GetTopN(10)
@@ -140,8 +140,8 @@ func TestBoundedCounter_Clear(t *testing.T) {
 	bc.Clear()
 
 	assert.Equal(t, 0, bc.Len())
-	assert.Equal(t, 0, bc.Get("foo"))
-	assert.Equal(t, 0, bc.Get("bar"))
+	assert.Equal(t, int64(0), bc.Get("foo"))
+	assert.Equal(t, int64(0), bc.Get("bar"))
 }
 
 func TestBoundedCounter_MaintainsTopItems(t *testing.T) {
@@ -167,9 +167,9 @@ func TestBoundedCounter_MaintainsTopItems(t *testing.T) {
 	// Verify top items are preserved
 	topItems := bc.GetTopN(10)
 	assert.Equal(t, "A", topItems[0].Key)
-	assert.Equal(t, 100, topItems[0].Count)
+	assert.Equal(t, int64(100), topItems[0].Count)
 	assert.Equal(t, "B", topItems[1].Key)
-	assert.Equal(t, 95, topItems[1].Count)
+	assert.Equal(t, int64(95), topItems[1].Count)
 }
 
 func TestBoundedCounter_ZeroCapacity(t *testing.T) {
@@ -180,7 +180,7 @@ func TestBoundedCounter_ZeroCapacity(t *testing.T) {
 
 	// Should not store anything
 	assert.Equal(t, 0, bc.Len())
-	assert.Equal(t, 0, bc.Get("foo"))
+	assert.Equal(t, int64(0), bc.Get("foo"))
 }
 
 func TestBoundedCounter_SingleCapacity(t *testing.T) {
@@ -188,11 +188,11 @@ func TestBoundedCounter_SingleCapacity(t *testing.T) {
 
 	bc.Increment("first")
 	assert.Equal(t, 1, bc.Len())
-	assert.Equal(t, 1, bc.Get("first"))
+	assert.Equal(t, int64(1), bc.Get("first"))
 
 	// Add second item - should replace first
 	bc.Increment("second")
 	assert.Equal(t, 1, bc.Len())
-	assert.Equal(t, 0, bc.Get("first"))
-	assert.Equal(t, 1, bc.Get("second"))
+	assert.Equal(t, int64(0), bc.Get("first"))
+	assert.Equal(t, int64(1), bc.Get("second"))
 }
